+++ coloraide/colors/_distance.py
@@ -1,7 +1,7 @@
 import math
 from .. import util
 
-G_CONST = math.pow(25, 7)
+G_CONST = 25 ** 7
 SUPPORTED = frozenset(["76", "2000", "cmc", "94"])
 
 
@@ -14,10 +14,7 @@ def distance_euclidean(color1, color2, space="lab", **kwargs):
     coords1 = util.no_nan(lab1.coords())
     coords2 = util.no_nan(lab2.coords())
 
-    total = 0
-    for i, coord in enumerate(coords1):
-        total += math.pow(coords2[i] - coord, 2)
-    return math.sqrt(total)
+    return math.sqrt(sum((x - y) ** 2.0 for x, y in zip(coords2, coords1)))
 
 
 def delta_e_76(color1, color2, **kwargs):
@@ -41,9 +38,9 @@ def delta_e_94(color1, color2, kl=1, k1=0.045, k2=0.015):
     lab2 = color2.convert("lab")
 
     l1, a1, b1 = util.no_nan(lab1.coords())
-    c1 = math.sqrt(math.pow(a1, 2) + math.pow(b1, 2))
+    c1 = math.sqrt(a1 ** 2 + b1 ** 2)
     l2, a2, b2 = util.no_nan(lab2.coords())
-    c2 = math.sqrt(math.pow(a2, 2) + math.pow(b2, 2))
+    c2 = math.sqrt(a2 ** 2 + b2 ** 2)
 
     dl = l1 - l2
     dc = c1 - c2
@@ -54,7 +51,7 @@ def delta_e_94(color1, color2, kl=1, k1=0.045, k2=0.015):
     # We never reference `dh` until the very end, and when we do, we square it
     # before using it, so we don't need the square root as described in the
     # algorithm. Instead we can just leave the result as is.
-    dh = math.pow(da, 2) + math.pow(db, 2) - math.pow(dc, 2)
+    dh = da ** 2 + db ** 2 - dc ** 2
 
     sl = 1
     sc = 1 + k1 * c1
@@ -64,10 +61,10 @@ def delta_e_94(color1, color2, kl=1, k1=0.045, k2=0.015):
     kh = 1
 
     return math.sqrt(
-        math.pow(dl / (kl * sl), 2) +
-        math.pow(dc / (kc * sc), 2) +
+        (dl / (kl * sl)) ** 2 +
+        (dc / (kc * sc)) ** 2 +
         # Square root just the denominator as `dh` is already squared.
-        dh / math.pow(kh * sh, 2)
+        dh / ((kh * sh) ** 2)
     )
 
 
@@ -82,9 +79,9 @@ def delta_e_cmc(color1, color2, l=2, c=1):
     lab2 = color2.convert("lab")
 
     l1, a1, b1 = util.no_nan(lab1.coords())
-    c1 = math.sqrt(math.pow(a1, 2) + math.pow(b1, 2))
+    c1 = math.sqrt(a1 ** 2 + b1 ** 2)
     l2, a2, b2 = util.no_nan(lab2.coords())
-    c2 = math.sqrt(math.pow(a2, 2) + math.pow(b2, 2))
+    c2 = math.sqrt(a2 ** 2 + b2 ** 2)
 
     dl = l1 - l2
     dc = c1 - c2
@@ -95,7 +92,7 @@ def delta_e_cmc(color1, color2, l=2, c=1):
     # We never reference `dh` until the very end, and when we do, we square it
     # before using it, so we don't need the square root as described in the
     # algorithm. Instead we can just leave the result as is.
-    dh = math.pow(da, 2) + math.pow(db, 2) - math.pow(dc, 2)
+    dh = da ** 2 + db ** 2 - dc ** 2
 
     if l1 < 16:
         sl = 0.511
@@ -115,16 +112,16 @@ def delta_e_cmc(color1, color2, l=2, c=1):
         t = 0.56 + abs(0.2 * math.cos((h1 + 168) * util.DEG2RAD))
     else:
         t = 0.36 + abs(0.4 * math.cos((h1 + 35) * util.DEG2RAD))
-    c1_4 = math.pow(c1, 4)
+    c1_4 = c1 ** 4
     f = math.sqrt(c1_4 / (c1_4 + 1900))
 
     sh = sc * ((f * t) + 1 - f)
 
     return math.sqrt(
-        math.pow(dl / (l * sl), 2) +
-        math.pow(dc / (c * sc), 2) +
+        (dl / (l * sl)) ** 2 +
+        (dc / (c * sc)) ** 2 +
         # Square root just the denominator as `dh` is already squared.
-        dh / math.pow(sh, 2)
+        dh / (sh ** 2)
     )
 
 
@@ -144,20 +141,20 @@ def delta_e_2000(color1, color2, kl=1, kc=1, kh=1, **kwargs):
     lab2 = color2.convert("lab")
 
     l1, a1, b1 = util.no_nan(lab1.coords())
-    c1 = math.sqrt(math.pow(a1, 2) + math.pow(b1, 2))
+    c1 = math.sqrt(a1 ** 2 + b1 ** 2)
     l2, a2, b2 = util.no_nan(lab2.coords())
-    c2 = math.sqrt(math.pow(a2, 2) + math.pow(b2, 2))
+    c2 = math.sqrt(a2 ** 2 + b2 ** 2)
 
     cm = (c1 + c2) / 2
 
-    c7 = math.pow(cm, 7)
+    c7 = cm ** 7
     g = 0.5 * (1 - math.sqrt(c7 / (c7 + G_CONST)))
 
     ap1 = (1 + g) * a1
     ap2 = (1 + g) * a2
 
-    cp1 = math.sqrt(math.pow(ap1, 2) + math.pow(b1, 2))
-    cp2 = math.sqrt(math.pow(ap2, 2) + math.pow(b2, 2))
+    cp1 = math.sqrt(ap1 ** 2 + b1 ** 2)
+    cp2 = math.sqrt(ap2 ** 2 + b2 ** 2)
 
     hp1 = 0 if (ap1 == 0 and b1 == 0) else math.atan2(b1, ap1)
     hp2 = 0 if (ap2 == 0 and b2 == 0) else math.atan2(b2, ap2)
@@ -202,20 +199,20 @@ def delta_e_2000(color1, color2, kl=1, kc=1, kh=1, **kwargs):
         (0.20 * math.cos(((4 * hpm) - 63) * util.DEG2RAD))
     )
 
-    dt = 30 * math.exp(-1 * math.pow(((hpm - 275) / 25), 2))
+    dt = 30 * math.exp(-1 * ((hpm - 275) / 25) ** 2)
 
-    cpm7 = math.pow(cpm, 7)
+    cpm7 = cpm ** 7
     rc = 2 * math.sqrt(cpm7 / (cpm7 + G_CONST))
-    l_temp = math.pow(lpm - 50, 2)
+    l_temp = (lpm - 50) ** 2
     sl = 1 + ((0.015 * l_temp) / math.sqrt(20 + l_temp))
     sc = 1 + 0.045 * cpm
     sh = 1 + 0.015 * cpm * t
     rt = -1 * math.sin(2 * dt * util.DEG2RAD) * rc
 
     return math.sqrt(
-        math.pow((dl / (kl * sl)), 2) +
-        math.pow((dc / (kc * sc)), 2) +
-        math.pow((dh / (kh * sh)), 2) +
+        (dl / (kl * sl)) ** 2 +
+        (dc / (kc * sc)) ** 2 +
+        (dh / (kh * sh)) ** 2 +
         rt * (dc / (kc * sc)) * (dh / (kh * sh))
     )