+++ coloraide/colors/srgb.py
@@ -52,7 +52,7 @@ def lin_srgb(rgb):
         if abs_i < 0.04045:
             result.append(i / 12.92)
         else:
-            result.append(math.copysign(math.pow((abs_i + 0.055) / 1.055, 2.4), i))
+            result.append(math.copysign(((abs_i + 0.055) / 1.055) ** 2.4, i))
     return result
 
 
@@ -68,7 +68,7 @@ def gam_srgb(rgb):
         # Mirror linear nature of algorithm on the negative axis
         abs_i = abs(i)
         if abs_i > 0.0031308:
-            result.append(math.copysign((1.055 * math.pow(abs_i, 1 / 2.4) - 0.055), i))
+            result.append(math.copysign((1.055 * abs_i ** (1 / 2.4) - 0.055), i))
         else:
             result.append(12.92 * i)
     return result